+++ src/cmake/testing.cmake
@@ -156,6 +156,7 @@ macro (oiio_add_all_tests)
                     diff
                     flip
                     dither dup-channels
+                    exif-usercomment
                     jpeg jpeg-corrupt jpeg-metadata
                     maketx oiiotool-maketx
                     misnamed-file